# api-docs.json
{"openapi":"3.0.1","info":{"title":"OpenAPI definition","version":"v0"},"servers":[{"url":"http://localhost:8086","description":"Generated server url"}],"paths":{"/cart/update":{"put":{"tags":["cart-controller"],"operationId":"updateCartQuantity","parameters":[{"name":"cartId","in":"query","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"quantity","in":"query","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Cart"}}}}}}},"/seller/register":{"post":{"tags":["auth-controller"],"operationId":"registerSeller","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserRegistrationRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/seller/products":{"post":{"tags":["product-controller"],"operationId":"addProduct","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Product"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/orders/place":{"post":{"tags":["order-controller"],"operationId":"placeOrder","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/OrderRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/cart/add":{"post":{"tags":["cart-controller"],"operationId":"addToCart","parameters":[{"name":"productId","in":"query","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"quantity","in":"query","required":true,"schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Cart"}}}}}}},"/buyer/register":{"post":{"tags":["auth-controller"],"operationId":"registerBuyer","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserRegistrationRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/authenticate":{"post":{"tags":["auth-controller"],"operationId":"createAuthenticationToken","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/JwtRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/admin/register":{"post":{"tags":["auth-controller"],"operationId":"registerAdmin","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserRegistrationRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/user":{"get":{"tags":["resource-controller"],"operationId":"userEndpoint","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/userId":{"get":{"tags":["auth-controller"],"operationId":"getUserId","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"integer","format":"int64"}}}}}}},"/seller/productsview":{"get":{"tags":["product-controller"],"operationId":"getSellerProducts","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/login":{"get":{"tags":["resource-controller"],"operationId":"loginEndpoint","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/cart/{userId}":{"get":{"tags":["cart-controller"],"operationId":"getCart","parameters":[{"name":"userId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Cart"}}}}}}}},"/buyer/products":{"get":{"tags":["product-controller"],"operationId":"getProducts","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Product"}}}}}}}},"/all":{"get":{"tags":["resource-controller"],"operationId":"allRolesEndpoint","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/admin":{"get":{"tags":["resource-controller"],"operationId":"adminEndpoint","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/seller/products/{id}":{"delete":{"tags":["product-controller"],"operationId":"deleteProduct","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/delete":{"delete":{"tags":["resource-controller"],"operationId":"deleteEndpoint","requestBody":{"content":{"application/json":{"schema":{"type":"string"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}}},"components":{"schemas":{"Cart":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"userId":{"type":"integer","format":"int64"},"productId":{"type":"integer","format":"int64"},"quantity":{"type":"integer","format":"int32"}}},"UserRegistrationRequest":{"type":"object","properties":{"username":{"type":"string"},"password":{"type":"string"},"role":{"type":"string","enum":["ADMIN","BUYER","SELLER"]}}},"Product":{"type":"object","properties":{"productId":{"type":"integer","format":"int64"},"productName":{"type":"string"},"productImg":{"type":"string"},"productDesc":{"type":"string"},"productPrice":{"type":"integer","format":"int32"},"productDiscount":{"type":"integer","format":"int32"},"productFeatures":{"type":"string"},"userId":{"type":"integer","format":"int64"}}},"OrderRequest":{"type":"object","properties":{"userId":{"type":"integer","format":"int64"},"productId":{"type":"integer","format":"int64"},"userMobileNumber":{"type":"string"},"userAddress":{"type":"string"}}},"JwtRequest":{"type":"object","properties":{"username":{"type":"string"},"password":{"type":"string"}}}}}}